Create ChartMain Component and put it inside Search Component
